Update dsrc package name (#127)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
This PR updates the package and node name for the v2x ros driver. In
order to represent the driver's functionality as being agnostic of
DSRC/CV2X the naming in the package and the variables defined for the
node are renamed to v2x radio.

Some variables which are being drawn from the j2735 spec are unchanged,
for example dsrc_id is still the official id field for messages in the
SAE J2735 spec.
The driver has previously been tested successfully with both DSRC and
CV2X radios. For this change it was tested on CV2X radios.

A missing change here is that sonarscanner needs to be updated for the
renamed repository. Currently the paths were updated so sonar scanner is
building the new code, but it is still writing to
usdot-fhwa-stol_CARMACohdaDsrcDriver. This change wasn't implemented
since a new repo needs to be configured manually. An issue was created
to make this change here
https://usdot-carma.atlassian.net/browse/ARPA-32

## Description

<!--- Describe your changes in detail -->

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key

<!-- e.g. CAR-123 -->
https://usdot-carma.atlassian.net/browse/ARPA-16

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
V2X communication tested between two vehicles using CV2X radios.
BSM messages seen to be communicated across the vehicles.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.

---------

Co-authored-by: John Stark <starkj@leidos.com>