Add optional flag to manage lifecycle transitions during launch (#133)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
Fixes launch file for v2x-ros-driver and adds optional flag to enable
lifecycle transition for v2x-ros-driver.
This change is required to allow users not using a lifecycle manager to
be able to transition the node to active state directly from the launch
file.
In the CARMA ecosystem this driver is used with carma-platform which
defines a lifecycle manager for the node. Having a lifecycle manager is
recommended to be able to gracefully activate, track failures and
gracefully shutdown the node, but adding the optional argument allows
users to use the driver in the absence of one as well.

## Description

<!--- Describe your changes in detail -->
This change set was tested locally using just the launch file for the
v2x_ros_driver node as well as on the Black Pacifica calling the
v2x-ros-driver node to launch through carma-platform's launch system.
The node lifecycle was confirmed to be "active" when the
enable_v2x_driver_lifecycle argument was set to true and "unconfigured"
with it set to false.

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key

<!-- e.g. CAR-123 -->
https://usdot-carma.atlassian.net/browse/ARC-178

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.