Fix parsing of received PSMs (#137)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

This PR introduces a fix to enable parsing of received PSMs. Added
additional parsing case to ensure a PSM is not accidentally classified
as a PSID, as it shares the same value as a BSM PSID (0x0020).

Additionally this PR adds the SensorDataSharingMessage (SDSM) to the
wave.json file so that received or meant-for-broadcast SDSMs will not be
discarded.

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key

ARC-372

## Motivation and Context

Parsing worked well with all messages except for received PSMs. PSMs
were getting skipped, but these changes verify that a PSM is not a PSID
and passes them as a valid message.

## How Has This Been Tested?

Tested with data collected from:
- Cohda MK6 OBU running RSU firmware, containing TIM, BSM, and PSM
without T-Headers
- Cohda MK6 OBU running OBU firmware, containing BSM, SPAT, MAP with
T-Headers.
- Kapsch OBU running OBU firmware, containing BSM, SPAT, MAP without
T-Headers

Also tested in White Pacifica with PSMs, TCRs, TCMs, MAP, and SPAT.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [X] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [X] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.

---------

Co-authored-by: chengyuan0124 <misheel.bayart@gmail.com>