modified:   docs/check-mk6-conf.sh (#151)
modified:   docs/mk6-rsu-obu-setup.sh

<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

Updated check-mk6-conf.sh and mk6-rsu-obu-setup.sh to include the 8002
and E0000017 PSIDs.

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key

<!-- e.g. CAR-123 -->

## Motivation and Context

This was needed because SPAT and MAP messages use those two PSIDs and
were not included in the previous configuration, causing those messages
to not be forwarded back to the V2X ROS Driver. These changes address
the issue.

## How Has This Been Tested?

Tested on the White Pacifica. Updated scripts were used to configure the
MK6 OBU. Received MAP and SPAT messages were then observed at the
inbound_binary rostopic.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.